+++ inquire/examples/confirm.rs
@@ -17,6 +17,7 @@ fn main() {
 
     let ans = Confirm {
         message: "Are you happy?",
+        starting_input: None,
         default: Some(false),
         placeholder: Some("si|no"),
         help_message: Some("It's alright if you're not"),